fix(agent): trim multi-line completion instead of discard.
